+++ source/toy_array.h
@@ -4,12 +4,12 @@
 
 //standard generic array
 typedef struct Toy_Array { //32 | 64 BITNESS
-	size_t capacity;       //4  | 4
-	size_t count;          //4  | 4
+	unsigned int capacity; //4  | 4
+	unsigned int count;    //4  | 4
 	char data[];           //-  | -
 } Toy_Array;               //8  | 8
 
-TOY_API Toy_Array* Toy_resizeArray(Toy_Array* array, size_t capacity);
+TOY_API Toy_Array* Toy_resizeArray(Toy_Array* array, unsigned int capacity);
 
 #define TOY_ALLOCATE_ARRAY(type, count) \
 	Toy_resizeArray(NULL, sizeof(type)*(count))
